+++ R/prediction_methods.R
@@ -137,6 +137,34 @@ prediction.nls <- function(model, data, ...) {
 
 #' @rdname prediction
 #' @export
+prediction.survreg <- function(model, data, type = c("response", "lp", "quantile", "uquantile"), ...) {
+    # setup data
+    if (missing(data)) {
+        if (!is.null(model[["call"]][["data"]])) {
+            data <- eval(model[["call"]][["data"]], parent.frame()) 
+        } else { 
+            data <- get_all_vars(model[["terms"]], data = model[["model"]])
+        }
+    }
+    
+    type <- match.arg(type)
+    
+    # extract predicted value at input value (value can only be 1 number)
+    pred <- predict(model, newdata = data, type = type, se.fit = TRUE, ...)
+    class(pred[["fit"]]) <- c("fit", "numeric")
+    class(pred[["se.fit"]]) <- c("se.fit", "numeric")
+    
+    # obs-x-2 data.frame of predictions
+    structure(list(fitted = pred[["fit"]], 
+                   se.fitted = pred[["se.fit"]]), 
+              class = c("prediction", "data.frame"), 
+              row.names = seq_len(length(pred[["fit"]])),
+              model.class = class(model),
+              type = type)
+}
+
+#' @rdname prediction
+#' @export
 prediction.coxph <- function(model, data, type = c("risk", "expected", "lp"), ...) {
     # setup data
     if (missing(data)) {